Incorporando alterações
