+++ src/main/java/br/com/zup/ClinicaGeneralMed/controllers/MedicoController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("medico")
+@RequestMapping("/medico")
 public class MedicoController {
     @Autowired
     private MedicoService medicoService;
