--- src/main/java/br/com/zup/ClinicaGeneralMed/dtos/Paciente.java
@@ -1,13 +1,0 @@
-package br.com.zup.ClinicaGeneralMed.dtos;
-
-
-public class Paciente extends Pessoa{
-
-    public Paciente() {
-
-    }
-
-}
-
-
-
