+++ src/main/java/br/com/zup/ClinicaGeneralMed/dtos/PacienteDTO.java
@@ -1,46 +1,12 @@
 package br.com.zup.ClinicaGeneralMed.dtos;
 
-public class PacienteDTO {
 
-    private String nome;
-    private String cpf;
-    private String endereco;
-    private String telefone;
+public class PacienteDTO extends Pessoa{
 
     public PacienteDTO() {
+
     }
 
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getEndereco() {
-        return endereco;
-    }
-
-    public void setEndereco(String endereco) {
-        this.endereco = endereco;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
 }
