--- src/main/java/br/com/zup/ClinicaGeneralMed/enums/Especialidades.java
@@ -1,9 +1,0 @@
-package br.com.zup.ClinicaGeneralMed.enums;
-
-public enum Especialidades {
-
-    CARDIOLOGIA,
-    PEDIATRA,
-    ORTOPEDISTA,
-    CLINICOGERAL,
-}
