+++ src/main/java/br/com/zup/ClinicaGeneralMed/services/PacienteService.java
@@ -1,6 +1,5 @@
 package br.com.zup.ClinicaGeneralMed.services;
 
-import br.com.zup.ClinicaGeneralMed.dtos.MedicoDTO;
 import br.com.zup.ClinicaGeneralMed.dtos.PacienteDTO;
 import org.springframework.stereotype.Service;
 
@@ -18,11 +17,13 @@ public class PacienteService {
                 throw new RuntimeException("Paciente já existe!");
             }
         }
+
         return paciente;
     }
 
     public PacienteDTO cadastrar(PacienteDTO paciente){
         this.pacientes.add(verificaPaciente(paciente));
+
         return paciente;
     }
